+++ packages/mint-components/src/components/sqm-rewards-table/columns/sqm-rewards-table-status-column.tsx
@@ -7,7 +7,7 @@ import { RewardTableColumn } from "./RewardTableColumn";
  * @uiName Reward Table Status Column
  * @validParents ["sqm-rewards-table"]
  * @exampleGroup Rewards
- * @example Reward Table Status Column - <sqm-rewards-table-status-column column-title="Status" status-text="{status, select, AVAILABLE {Available} CANCELLED {Cancelled} PENDING {Pending} EXPIRED {Expired} REDEEMED {Redeemed} other {Not available} }" expiry-text="Expires on " pending-us-tax="W-9 required" pending-scheduled="Until" pending-unhandled="Fulfillment error"></sqm-rewards-table-status-column>
+ * @example Reward Table Status Column - <sqm-rewards-table-status-column column-title="Status" status-text="{status, select, AVAILABLE {Available} CANCELLED {Cancelled} PENDING {Pending} PENDING_REVIEW {Pending} DENIED {Denied} EXPIRED {Expired} REDEEMED {Redeemed} other {Not available} }" expiry-text="Expires on " pending-us-tax="W-9 required" pending-scheduled="Until" pending-unhandled="Fulfillment error" pending-review-text="Awaiting review" denied-text="Detected self-referral"></sqm-rewards-table-status-column>
  */
 @Component({
   tag: "sqm-rewards-table-status-column",
@@ -24,7 +24,7 @@ export class RewardTableStatusColumn implements RewardTableColumn {
    * @uiWidget textArea
    */
   @Prop() statusText: string =
-    "{status, select, AVAILABLE {Available} CANCELLED {Cancelled} PENDING {Pending} EXPIRED {Expired} REDEEMED {Redeemed} other {Not available} }";
+    "{status, select, AVAILABLE {Available} CANCELLED {Cancelled} PENDING {Pending} PENDING_REVIEW {Pending} EXPIRED {Expired} REDEEMED {Redeemed} DENIED {Denied} other {Not available} }";
 
   /**
    * Text shown before the date of an expiring reward.
@@ -54,6 +54,20 @@ export class RewardTableStatusColumn implements RewardTableColumn {
    */
   @Prop() pendingUnhandled: string = "Fulfillment error";
 
+  /**
+   *  Displayed when flagged for fraud.
+   *
+   * @uiName Pending review text
+   */
+  @Prop() pendingReviewText: string = "Awaiting review";
+
+  /**
+   *  Displayed when denied for fraud.
+   *
+   * @uiName Denied text
+   */
+  @Prop() deniedText: string = "Detected self-referral";
+
   constructor() {
     withHooks(this);
   }
@@ -69,6 +83,8 @@ export class RewardTableStatusColumn implements RewardTableColumn {
         pendingScheduled={this.pendingScheduled}
         pendingUsTax={this.pendingUsTax}
         pendingUnhandled={this.pendingUnhandled}
+        pendingReviewText={this.pendingReviewText}
+        deniedText={this.deniedText}
         locale={locale}
       ></sqm-rewards-table-status-cell>
     );
@@ -84,6 +100,7 @@ export class RewardTableStatusColumn implements RewardTableColumn {
       this.columnTitle,
       this.statusText,
       this.expiryText,
+      this.pendingReviewText,
       this.pendingScheduled,
       this.pendingUsTax,
       this.pendingUnhandled,